feat(backend/api): Рефакторинг
