+++ backend/src/api/v1/login.rs
@@ -1,0 +1,29 @@
+use crate::api::{auth, payloads::*};
+use axum::{http::StatusCode, response::IntoResponse, Json};
+
+pub async fn login(
+    mut auth_session: auth::AuthSession,
+    Json(payload): Json<LoginPayload>,
+) -> impl IntoResponse {
+    let creds = payload.clone().into();
+    let user = match auth_session.authenticate(creds).await {
+        Ok(Some(user)) => user,
+        Ok(None) => return StatusCode::UNAUTHORIZED.into_response(),
+        Err(_) => return StatusCode::INTERNAL_SERVER_ERROR.into_response(),
+    };
+
+    if let Err(err) = auth_session.login(&user).await {
+        tracing::error!("Failed to login user {user}: {err}", user = user.0.name);
+        return StatusCode::INTERNAL_SERVER_ERROR.into_response();
+    }
+
+    Json(user.0).into_response()
+}
+
+pub async fn logout(mut auth_session: auth::AuthSession) -> StatusCode {
+    match auth_session.logout().await {
+        Ok(Some(_)) => StatusCode::OK,
+        Ok(None) => StatusCode::UNAUTHORIZED,
+        Err(_) => StatusCode::INTERNAL_SERVER_ERROR,
+    }
+}
