+++ backend/src/domain/mod.rs
@@ -5,5 +5,6 @@ mod participant;
 pub use adult::{Adult, AdultId, AdultRole};
 pub use operations::{Order, Sort};
 pub use participant::{
-    AnonymousParticipant, Participant, ParticipantId, ParticipantInfo, ParticipantRate,
+    JuryParticipant, Participant, ParticipantAnswers, ParticipantCode, ParticipantId,
+    ParticipantInfo, ParticipantRate,
 };
